Split into separate files + add xmldocs
